Improve errors when `tensorstore.open` is called with invalid spec
The `tensorstore.open` method takes a `tensorstore.Spec` object, but
`tensorstore.Spec` supports implicit conversion from JSON.

Previously, this implicit conversion was provided using
`pybind11::implicitly_convertible`, but that swallows conversion
errors and simply indicates that overload resolution failed.

While that behavior is useful for allowing a different overload to be
selected, in practice we don't need that behavior and it is more
important to return a more informative error to the user.

PiperOrigin-RevId: 307142759
Change-Id: I743db2b299b7cc1d73934157bf9142d4c0f9bbaa