Factor out AggregateWritebackCache
Factor out logic for handling AsyncStorageBackedCache "write states"
that consist of multiple independent atomic changes to the same state
from MetadataCache in kvs_backed_chunk_driver and
ShardedKeyValueStoreWriteCache in uint64_sharded_key_value_store.

This allows the same logic to be shared for single image file drivers
to be added in a later change.

PiperOrigin-RevId: 312332507
Change-Id: Ic30fb6cd5b09f9811bb76dd165a2fc7436a8ba18