chunk_cache: Support unconditional writes of partial chunks
Previously, chunk_cache would only consider a chunk fully overwritten
(and thereby eligible for unconditional writeback) if the entire chunk
was overwitten, even if the chunk extends beyond the bounds of the
TensorStore.  For resizable dimensions, this behavior is correct to
avoid race conditions, but for non-resizable dimensions, this leads to
unnecessary conditional writebacks, which for sharded
neuroglancer_precomputed volumes can be costly, as reported in
https://github.com/google/tensorstore/issues/5.

This does not fix by itself
https://github.com/google/tensorstore/issues/5 but allows a later
change to address that issue to also handle partial chunks at the
upper bound of a neuroglancer_precomputed volume.

PiperOrigin-RevId: 312869823
Change-Id: Icc94046105669fff763ab2a68ef31163c404a796