uint64_sharded_key_value_store: Support unconditional full shard writes
Previously, all writes to a `uint64_sharded_key_value_store` resulted
in conditional writes to the underlying KeyValueStore because there
was no mechanism for determining that all possible chunks within the
shard were being written unconditionally.  Consequently, even in code
carefully designed to perform shard-aligned writes, if due to a
previous failure there was an existing shard file to be overwritten,
writeback would involve an extra unnecessary read and an extra
unnecessary write of the entire shard.

This change adds a mechanism by which the "neuroglancer_precomputed"
driver can indicate to the `Uint64ShardedKeyValueStore` the number of
chunks in each shard, and modifies `Uint64ShardedKeyValueStore` to
perform an unconditional write when all chunks in an entire shard are
overwritten unconditionally.

This partially addresses
https://github.com/google/tensorstore/issues/5 but does not fix it
completely, because there is still a race condition in which writeback
of a shard starts before all chunk writes buffered in memory have been
processed.  That will be addressed by a later change that adds
transaction support.

PiperOrigin-RevId: 312869870
Change-Id: I6109f894ab49ea1d26979564ebe611a79e7ee41d