Remove the global operator| overload.
The pipeline operator is intended to make tensorstore easier to use. The current implementation uses a global operator| to match on  Result<T> or functions having the signature Result<U>(T).

This introduces a lot of compile-time overhead matching all operator | calls, and, in addition, global operator overloading is generally frowned upon, particularly by the Google style guide.  In addition, the global overload introduces a lot of language-level ambuigity by attempting to match on types like this:

Result<int> foo(int);
... later ...

auto bar = 1 | foo;

While the natural reading of bar being an int, the global operator| overload may be invoked, turning this into a function call.

I think that it's more clear and just as useful to eliminate the global overload and just supply operator| on those types where the support is most worthwhile. This includes the types that are currently subject to indexing operations which benefit from being chained.

PiperOrigin-RevId: 324234175
Change-Id: I6f083f189622a7f20472a0d0ef70d9c118249083