Fix CURLE_SEND_FAIL_REWIND errors when using GCS driver
With the change to use absl::Cord rather than std::string, we
neglected to specify a CURL_SEEKFUNCTION, which Curl requires in order
to internally resend requests, and also did not treat
CURLE_SEND_FAIL_REWIND as a retryable error, which is triggered when
Curl would like to retry a request but no CURL_SEEKFUNCTION has been
specified.

Curl attempts to retry a request if the server closes the connection
just as curl is attempting to reuse it for another request.

This change defines a CURL_SEEKFUNCTION in order to avoid this problem.

PiperOrigin-RevId: 325544401
Change-Id: I042d38378c7c2430bdf562e46476364fe0e93723