Allow WriteChunk::EndWrite to report copy errors
Previously, the WriteChunk::EndWrite method had no way to report
errors copying the data; any such errors had to be reported by the
NDIterable itself.  The returned Future was required to be the future
associated with the transaction, and was ignored in the case of an
explicit transaction, because it was assumed to be identical.

With this change, WriteChunk::EndWrite gains an extra Status return
which can be used for copy errors; this functionality is needed by the
"json" driver added in a subsequent change.

PiperOrigin-RevId: 339282764
Change-Id: I9dc3b2fd3efa7ef13b50687cf80b45d321290047