Retry HTTP requests in case of CURLE_PARTIAL_FILE
When under heavy load, Google Cloud Storage sometimes terminates the
response early, resulting in `CURLE_PARTIAL_FILE` or `CURLE_HTTP2`,
particularly for large requests.  Currently, the request is retried
from the beginning in all cases.  It might be better to retry with a
byte range request from where the response ended, but this would have
to be done carefully to avoid data corruption.

PiperOrigin-RevId: 339283530
Change-Id: If5815fd78272efce2133b1b8ad90a9b171049ea0