Fix read/write mode selection in tensorstore.open
Previously, specifying neither `read` nor `write` resulted in
`ReadWriteMode::read_write` being used to open the TensorStore, which
fails if both reading and writing are not supported.

With this change, specifying neither `read` nor `write` results in
`ReadWriteMode::dynamic`, which results in the TensorStore being
opened in read, write, or read/write mode depending on what is
supported by the driver.

PiperOrigin-RevId: 344873479
Change-Id: I01f8ca0e3fd9d40d3bb96a3df6869051acc25757