Change n5 driver to always write full chunks
The n5 format allows each chunk to specify a shape in its header.  Previously, the n5 driver would write chunks that are partially outside the bounds specified in the metadata as "partial" chunks containing only the in-bounds portions, for consistency with some existing n5 implementations.

However, the chunk encoding logic always used the initial bounds when the dataset was opened, rather than the "current" bounds.  This led to corrupted chunk encoding or crashes in the case of certain resize operations.  Furthermore, it is not even possible to determine the bounds in a race-free manner.

For that reason, this commit changes the n5 driver to always just write full chunks, which simplifies the code, avoids corrupted chunks, and avoids the possibility of this race.

PiperOrigin-RevId: 352895570
Change-Id: I62087a9b7c35ee69da30c0a0afe708ffbfe36f74