Avoid defining pybind11 type_caster specializations for pybind11-wrapped classes
Previously, for `Spec`, `DimensionSelection`, and `DataType`
TensorStore defined type_caster specializations that conveniently
enabled additional implicit conversions.

However, due to the way the pybind11 type casting machinery works,
this is very error-prone because if any translation unit depends on
one of those conversions but fails to include the header that defines
the specialization, there is no compile-time error but there is an ODR
violation that in practice leads to mysterious runtime failures.
While there currently are not any such bugs, they have occurred
previously.

This commit replaces those specializations with type_caster
specializations for special non-pybind11-wrapped `*Like` types, which
makes explicit the fact that a conversion is allowed.

This is much less error prone, because for types that are not
pybind11-wrapped classes, failing to include the appropriate
type_caster specialization also does not lead to a compile-time
error (unfortunately), but at least leads to a guaranteed and obvious
run time error.

PiperOrigin-RevId: 355752191
Change-Id: Ia344a6bf57f277265de8825b02b4760c43173f2b