Fix lifetime extension issue with TENSORSTORE_RETURN_IF_ERROR
Previously, TENSORSTORE_RETURN_IF_ERROR was implemented over multiple
statements in such a way that if the expression evaluated to a
reference to a temporary constructed as part of the expression, an
invalid reference would be used.  This could occur in cases like:

    Future<void> Func();
    TENSORSTORE_RETURN_IF_ERROR(Func().result());

This problem is now fixed by making a copy of the absl::Status in the
same statement that evaluates the expression.

PiperOrigin-RevId: 356057718
Change-Id: I7fe9d4bb19ae99fc9b3c1119ef5f740836c4e6b8