Remove error-prone default GetStatus overload
Previously, there was a default overload of `GetStatus(T)` that just
returned `OkStatus()`.  This worked okay when original `GetStatus`
only was intended for `Result` and `Status`, but now that it also
supports `Future` and similar types, it is error prone since
forgetting to provide a `GetStatus` overload just means any errors are
silently ignored.

PiperOrigin-RevId: 357860592
Change-Id: I288c649f915bc8a7ed0cbcaad4356b0d23859fc9