Consistently set IncludeDefaults{false} in JSON serialization
Previously, the default JSON serialization included default values in
some cases but not others.  In general it does not make a lot of sense
to include default values since they clutter the output with often
redundant information.

With this commit, default values are consistently excluded by
default (but the behavior can still be overridden by
`IncludeDefaults{true}`).  Additionally, this replaces the `bool
DisallowIncludeDefaults` parameter to the `DefaultValue` JSON binder
with the more flexible (and self-documenting) IncludeDefaultsPolicy
enum parameter.

PiperOrigin-RevId: 365963835
Change-Id: I19ba235ab993ed1077c8d1715085b2df67fea18f