Improved JSON schema documentation generation
This commit changes both how schemas are cross-referenced, and also
how they are displayed.

Previously, JSON schemas were assigned full (synthetic) URLs as
identifiers, which follows standard practice for JSON schemas but made
them cumbersome to specify and unusable for display in the
documentation.

The custom json-schema directive always imported schemas based on a
file path, and cross-references between schemas were not handled very
well: it was necessary to specify a manual mapping from ids to paths
in the configuration.

Additionally, the json-schema directive was not properly integrated
into Sphinx's system of "domains" and "objects" (which provide support
for modules/classes/functions in various programming languages).

With this commit, shorter paths are used as JSON schema identifiers,
in order to allow those ids to be displayed in the documentation.

This commit also adds support for a `json` Sphinx domain with a
`json:schema` directive that behaves similarly to the old
`json-schema` directive.  JSON schemas are now rendered similarly to
other Sphinx "objects" like Python classes.  Schemas are also now
present in the Sphinx object search index.

Schemas are always referenced by their id rather than file path, and
are are discovered automatically by walking the source tree, which
avoids the need to manually specify the mapping from ids to paths.

PiperOrigin-RevId: 366342524
Change-Id: Ia78537a3bc5fff87e8e50011342e0a9da058b0de