Fix documentation of IndexTransform.output
Changes to jsonschema_sphinx.py accidentally broke support for arrays
of objects, where the object type is described inline (without a
separate schema identifier) rather than by $ref.

This commit fixes that, but also makes OutputIndexMap its own named
schema, since that term is used in the C++ and Python APIs.

Additionally, this commit adds a clarifying note to the IndexTransform
documentation regarding the interpretation of the input and output
spaces.

PiperOrigin-RevId: 377708176
Change-Id: Ib04cf4c94125bd905ae06f6959f2ee3ab78443d8