Fix support for HTTP POST requests with empty body
Previously, the default HTTP method was set based on the presence of a
body, but this led to incorrect behavior in the case of a POST request
with an empty body.

In particular, writing empty values to GCS did not work.

With this change, the HTTP method is always specified explicitly.

PiperOrigin-RevId: 381551831
Change-Id: I809a6beac3797de83c3b4996a3e412ef84621d80