Add Schema as a Spec / Open option
This integrates Schema into the driver framework, and adds interfaces
for obtaining the schema from an open TensorStore and from a Spec, and
for setting Schema constraints on a Spec and when opening.

This also adds implementations of these schema interfaces for the
array, cast, downsample, and json drivers.  The implementations for
the zarr, n5, and neuroglancer_precomputed drivers are added in
subsequent commits.

Since the Schema can specify a rank constraint, IndexTransformSpec
would be redundant and has been removed.

PiperOrigin-RevId: 384997240
Change-Id: Iaab7271654d07fca6661356471b8b710caf14371