Add AreArraysSameValueEqual function for handling -0.0 and NaN
The normal array equalty in TensorStore, when used on floating point
types, follows equality comparison rules of the floating point type,
meaning that -0.0 and +0.0 compare equal, and NaN compares unequal to
itself.

In some cases, such as when comparing fill values, it is important to
treat -0.0 unequal from +0.0, and treat NaN equal to NaN.  This is
considered "same value" equality.  Note that this is still not the
same as bitwise equality, because there are multiple possible bit
representations for NaN, all of which are considered "same value
equal".

The "same value equality" terminology was borrowed from the JavaScript
specification (Object.is):
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Equality_comparisons_and_sameness

In the Julia programming language, the operation is called "isequal".

PiperOrigin-RevId: 384997711
Change-Id: I8d9bd185b2408ae2d4c2bef8a5adee20f01c5731