Add support for zarr dimension_separator metadata field
Historically zarr has named chunk files using dots by default,
e.g. "1.2.3", but has also supported slashes, e.g. "1/2/3".  However,
the choice was not indicated in the .zarray metadata file, and instead
needed to be specified out-of-band.  In TensorStore, that was
accomplished using the "key_encoding" spec member.

Zarr recently added a dimension_separator field to the metadata
format:

https://github.com/zarr-developers/zarr-python/pull/715

This commit adds support for that field.  Note that previously
TensorStore did not accept .zarray files that contain unknown
members, meaning that reading arrays with the dimension_separator
field would fail.

The existing "key_encoding" spec member is still supported, but
deprecated in favor of specifying "dimension_separator" as part of the
"metadata" field.  When creating a new array, TensorStore will always
write the "dimension_separator" field, in order to allow the array to
be re-opened without specifying the "dimension_separator".

Additionally, this change also makes TensorStore accept (and preserve)
extra .zarray members, as the zarr spec has now also changed to
require that.

PiperOrigin-RevId: 384998911
Change-Id: I485141b7786084d57c28005e3076f70462cd5b9e