Eliminate separate BoundSpec types
Previously, the TensorStore C++ API included various "Spec" types,
which contained "unbound" Context::ResourceSpec objects, and
corresponding "BoundSpec" types, which contained "bound"
Context::Resource objects.

For example:

Context::ResourceSpec<Provider> and Context::Resource<Provider>
KeyValueStore::Spec::Ptr and KeyValueStore::Spec::Bound::Ptr
Driver::Spec::Ptr and Driver::BoundSpec::Ptr

When defining KeyValueStore and TensorStore drivers, the spec data
would be defined using a SpecT<MaybeBound> class template, in order to
define both variants at the same time.

This had the advantage of providing static typing of the context
binding state, but had a number of disadvantages:

- Proliferation of types templated on `MaybeBound`

- Lack of uniformity between the interfaces and supported operations
for the Spec and BoundSpec types.  For example:

  - Driver::Spec is accessed via the public
  tensorstore::Spec type, but there was no public interface for
  Driver::BoundSpec.

  - Conversion operations (e.g. setting schema constraints, changing
  open mode) could only be applied to Driver::Spec, not to
  Driver::BoundSpec.

- Additional functionality like concatenation views and multiscale
representation to be added in the future will exacerbate these issues

With this commit, `Context::Resource<Provider>` is changed to be a
variant type that can hold either a context resource spec or a context
resource.  That doesn't significantly sacrifice type safety since
`Context::Resource` could already be null (which was not valid for
most purposes).  Being in the resource spec state when a resource is
expected is just another invalid state.

This allows the separate BoundSpec types to be eliminated.

It is now possible in principle for a single Spec object to contain
both bound and unbound context resources.  Any unbound resources can
be bound by calling `Spec::BindContext`; the already-bound resources
remain the same, and any unbound resources are resolved.  It is not
expected that this is particularly useful, but it is not particularly
harmful either.  In practice there is currently no way to create such
a Spec, but that is likely to be possible in the future.  The Python
context pickling support has been updated to properly support
partially-bound specs.

This commit also includes several related changes:

- Previously, under the default option of `IncludeDefaults{false}`,
the following JSON representation:

{"driver": "memory", "context": {"memory_key_value_store": {}}}

would be converted back to just:

{"driver": "memory"}

However, this violates the intended invariant that
`IncludeDefaults{false}` should not affect the meaning, because with
the "context" specification, a new "memory_key_value_store" resource
is created, while without the "context" specification, the
"memory_key_value_store" resource will be obtained from the parent
context.

That is now fixed: the "context" specification is now included, to
properly represent the Spec.

- The tensorstore::IncludeContext JSON serialization option (and the
corresponding Python include_context option) is no longer supported.
Instead, to obtain the JSON representation without context resources,
first remove the context resources from the Spec itself, via the new
`strip_context` option.  Previously, context resources could only be
stripped by round tripping through JSON, which was unfortunate and
confusing.

PiperOrigin-RevId: 388579560
Change-Id: I8c61fc2d9b846c055e09e164bfe0c9186343f609