Prevent zero-element index arrays in index transform representation
The index array output index map representation supports index arrays
that are indexed by any subset of the full set of input dimensions.
Input dimensions included in the subset are indicated by a non-zero
byte_stride.

Index transforms also support input dimensions with a size of 0.
Furthermore, a dimension with implicit bounds can be resized to 0,
then resized back to a non-zero size, and the output index maps must
be preserved.

Previously, it was possible to obtain an index transform with an
explicit dimension of size 0 containing an index array output index
map.  In this case, no positions within the index array may be
accessed.  However, if the byte_stride of the index array for a
zero-size dimension was set to 0, then instead several code paths
treated the index array as being non-empty and simply not indexed by
the zero-size dimension.

With this change, we ensure that index arrays maps are not permitted
in index transforms that have any explicit zero-size dimension.  Since
such a transform has no valid input positions, it is not necessary to
preserve the output index maps, and all index array maps are converted
to constant maps with an offset of 0.

This bug was caught by Address Sanitizer.

PiperOrigin-RevId: 389053249
Change-Id: I4d721d6adad1efc849209413520248b98195e5b7