Fix out-of-bounds read in neuroglancer_compressed_segmentation decoding
Previously, if a block with `encoded_bits == 0` (i.e. only a single
distinct value) was indicated in the header as having its encoded
table indices starting at the end of the buffer (which is valid since
the encoded table indices are length 0), `DecodeBlock` would read 4
bytes starting at the end of the buffer (the result of the read was
ignored due to a shift operation).

This representation could be generated by the normal encoder in the
case where the last block contains only a single distinct value, and a
previous block also contained just that same distinct value (such that
a new label table would not have to be written).

This problem was caught by Address Sanitizer.

PiperOrigin-RevId: 389053758
Change-Id: Ieaa94144e54a3e0e390eadf356a8fefe16e865b5