Avoid racy debug assertions for Transaction
This removes the debug-only assertions on commit_state_ that were done
without acquiring the mutex, which triggered thread sanitizer (tsan)
warnings.

In practice these assertions were safe as long as reading or writing
the `commit_state_` would not tear (i.e. provide "relaxed" load/store
memory order), but as they are not likely to be particularly important
debugging checks they are removed to avoid the tsan warning.

PiperOrigin-RevId: 390466450
Change-Id: I592c2336ba263385a257767705d6a13fc5e1a80b