Rename KeyValueStore -> KvStore and add path abstraction
Renames KeyValueStore -> kvstore::Driver, KeyValueStoreSpec ->
kvstore::DriverSpec.

The new KvStore class combine a DriverPtr with a string path.  It
supports Read/Write/DeleteRange/List operations that simply forward to
the same method of the KeyValueStore, but where the specified paths
are relative to the `path`.

In practice virtually all uses of a key-value store also use a path.

Likewise, the new kvstore::Spec class combines a DriverSpecPtr with a
string path.

DriverSpecPtr is no longer directly convertable to/from JSON.
Instead, kvstore::Spec is used, which has an additional
"path" member in the JSON representation for specifying the path.  For
the "file" driver, which already had a "path" member, the internal
root path that was previously stored by the driver has been replaced
by the path stored within the KvStore.

PiperOrigin-RevId: 393205452
Change-Id: Ib5b1927c4cac8363967800918bbccb9dd41c1f69