Refactor Python binding definitions to avoid order dependence
In order for pybind11 to show the correct type names in its generated
function signatures, all of the parameter types must be registered
before the function itself is registered.

Previously, that was handled by carefully choosing the order in which
classes and their methods.

With this change, functions and methods are instead deferred to an
executor that runs immediately after all of the classes are
registered.

PiperOrigin-RevId: 395366435
Change-Id: I65ab0017379077919ae15ac4d119faa1ad0099f2