Consolidate index_space:{index_transform_builder,index_transform,json}
This is needed to avoid a circular dependency when serialization
support is added in a subsequent commit: we want serialization to be
provided by index_transform.h, but serialization depends on json
conversion.  Consequently, json conversion must be available in the
same build target as index_transform.h.

The new dimension_labels target is introduced to avoid a circular
dependency with internal:dimension_indexed_json_binder.

PiperOrigin-RevId: 396070209
Change-Id: Ie24c9aef5ecf4497de21adce0642c33a97ce7633