Improve interoperability of tensorstore::Future with Python asyncio
This adds support for encoding arbitrary Python exceptions as an
`absl::Status` (pickled representation is stored as an additional
payload value), and for creating a tensorstore::Future from a
coroutine.

This also partially mitigates crashes due to
https://bugs.python.org/issue42969: TensorStore will itself no longer
cause crashes when Python is finalizing (last stage of exiting) due to
attempting to acquire the GIL from other threads in response to
asynchronous notifications.  However, if a Python daemon thread is in
the middle of a call to tensorstore while the interpreter is
finalizing, a crash can still occur.  Users can avoid that by using an
`atexit` handler to ensure daemon threads exit before finalization.

PiperOrigin-RevId: 398877103
Change-Id: Ice0abb5198f6f8de2e2652215876757f139c5152