Add pybind11_cc_googletest_test rule
This defines a C++ test target that runs a C++ GoogleTest suite but
can make use of pybind11 and the Python C API.

Previously, the Python `index_test` target did not actually call any
Python APIs, but it depended on the `index` library target that did
depend on pybind11.  In some cases, the references to Python APIs were
not fully optimized out, and a link error would result.

PiperOrigin-RevId: 398877309
Change-Id: I6007add57caf30d1110b6cbdbde0a08126f12eee