Fix ProtobufSerializer::Decode() and update LimitingReader usage.
Follow LimitingReader changes in
https://github.com/google/riegeli/commit/f872372c10a8ef56c92880ff0c32c71dd5f4a04d:
use LimitingReaderBase::Options::set_exact_length() to specify the limit
relatively to the current position (this was incorrect, the size was interpreted
was a position), and to verify that the source is long enough.

Cosmetics: let ParseFromReader() own a temporary LimitingReader to avoid having
to explicitly close it.
PiperOrigin-RevId: 401310831
Change-Id: I55a96e9516d1efe1bb59f64b8d9239a9801a40b1