Refactor TensorStore DriverSpec interface to reduce indirection
This removes additional indirection from internal::DriverSpec that is
no longer necessary after the removal of separate BoundSpec types.

Previously, individual drivers would define a single `DerivedDriver`
class that inherits from the `internal::RegisteredDriver` CRTP base.
The `DerivedDriver` type was required to define a nested `SpecData`
type and various static member functions that perform operations on
`SpecData`.  The `RegisteredDriver` class defined a corresponding
derived `DriverSpec` class with a `SpecData` object as a data member
and with virtual methods implemented in terms of the static member
functions defined by `DerivedDriver`.

With this change, drivers implementations define a `DerivedDriverSpec`
class that inherits from the `RegisteredDriverSpec` CRTP base, and
define all virtual member functions, other than `Clone`, manually.

In addition to removing an unnecessary layer of indirection, this
change also better decouples the DriverSpec from the Driver, which is
needed by a subsequent change that introduces a `MultiscaleDriver`
type.

PiperOrigin-RevId: 410602320
Change-Id: Ibe042e10926ac38af0aca688b885c5395ce8ef81