Add support for applying an IndexTransform or DimExpression to an IndexDomain
While in most cases it would be sufficient to just apply the
DimExpression or IndexTransform to the underlying TransformRep of the
IndexDomain in exactly the same way as if it were being applied to an
IndexTransform, in some cases that could lead to spurious errors:

The `IndexTransform::domain` method just returns a pointer to the same
TransformRep as the IndexTransform.  If we then naively apply a
DimExpression or compose with another IndexTransform, the operation
will also apply to the output index maps, and the transformed "domain"
will still contain the updated output index maps.  Not only does add
additional cost, computing the updated output index maps may lead to
an out-of-range or integer overflow error.  To avoid that, various
index transform operations gain a `domain_only` parameter in their
internal interface, which if set indicates that the output index maps
are to be ignored.

This also moves IndexDomain to a new index_domain.h header since
index_transform.h had grown quite large.

IndexDomain now stores a TransformRep pointer directly, rather than an
IndexTransform, which eliminates the circular dependency between
IndexDomain and IndexTransform.

PiperOrigin-RevId: 410604034
Change-Id: I10d29ef7fa5472b0a529769584bb8dd6954314df