Release GIL during pickling and unpickling
Previously, the GIL was held during the entire serialization and
deserialization process.  For most types that caused no problems, but
unpickling `tensorstore.TensorStore` blocks synchronously in a call to
`Future::result` while waiting for the TensorStore to open.  That is
unsafe since a Future callback may attempt to acquire the GIL.

This fixes the Windows Github actions build; for some reason the
deadlock only occurred intermittently with the Windows Github actions
build.

PiperOrigin-RevId: 414090492
Change-Id: I243697c99206d973be488d02313638cbae4c38ec