Refactor kvstore::DriverSpec interface to reduce indirection
This brings the kvstore driver interface more in line with the main
TensorStore driver interface.

KvStore drivers now directly define a `DerivedSpec` class that
inherits from the kvstore::RegisteredDriverSpec CRTP base class, and
define `DoOpen` directly, rather than using the `DriverOpenState`
class which has now been eliminated.

This gives KvStore drives the flexibility to directly override virtual
methods of `kvstore::DriverSpec`, which is needed for the URL support
to be added in a subsequent commit.

PiperOrigin-RevId: 419502213
Change-Id: I62facf98290ef69dbb9a52e7d5ad9f60d2c014d0