Ensure zarr driver stores all-zero chunks if fill_value is unspecified
Previously, the zarr driver treated an unspecified fill
value (i.e. `null`) identically to a fill value of 0.  In particular,
it would still apply the same write optimization used for a fill value
of 0, where a chunk that is entirely equal to the fill value is
deleted.

When purely using tensorstore this causes no problems.  However, this
leads to an incompatibility with zarr-python, where if the fill_value
is unspecified, reading a non-existent chunk returns arbitrary data
that may have been left in memory.  Consequently, if the fill value is
unspecified, and tensorstore is used to explicitly write zeros, and
then the array is read using zarr, the read result will contain
arbitrary values from memory rather than the zeros that are expected.

Fixes https://github.com/google/tensorstore/issues/28.

PiperOrigin-RevId: 423226264
Change-Id: I3dd8f140ba37acd85b90cb5669f020ae5a33cfde