Initial "stack" tensorstore driver.
The "stack" driver composes a TensorStore from a set of "layers", where each layer includes a transform that describes its position in the TensorStore.

The implementation constructs a non-uniform grid mapping each grid-space to a given layer. Layers are opened on demand (uncached) when Read or Write is called. Bounds are set by the layer description, and Resizing is not permitted.

PiperOrigin-RevId: 425123183
Change-Id: Iba796dde757c39cbb0e2379213af87c8d2569748