Refactor estimate_heap_usage
This makes estimate_heap_usage work similarly to serialization,
and garbage collection.

Previously, due to the default implementation that returned 0, many
types were silently using incorrect heap usage estimates, and there
was also a risk of ODR violations.

With this change, `internal::HeapUsageEstimator` must be specialized
for each type, and a missing specialization leads to a compile-time
error.

PiperOrigin-RevId: 431197345
Change-Id: I914aa74b8bacfce70695bff632d73a4bdf4b0a15