Save cache_identifier for open kvstore cache
Previously, upon destruction of a kvstore, its cache key was
regenerated in order to check whether it is present in the open
kvstore cache.  With this change, the cache identifier is saved which
is more efficient and avoids the potential for use-after-free if the
cache identifier changes for some reason.

PiperOrigin-RevId: 431197426
Change-Id: I3bbf90fc0a94e14a52f501dba4ebaf461791c982