Change AsyncWriteArray to use copy-on-write
Previously AsyncWriteArray (used by ChunkCache) required locking the
entire AsyncWriteArray object while accessing the WritebackData.  As a
consequence, in the common case where the data already has the correct
memory layout, an unnecessary copy is still required during writeback.

With this change, `AsyncWriteArray` uses copy-on-write, which allows
any data structures used by the writeback implementation to safely
hang onto a reference to the same backing array used by
AsyncWriteArray.

As the extra lock parameter to the AsyncCache::ApplyReceiver existed
solely to support AsyncWriteArray, it is now removed which simplifies
the interface.

PiperOrigin-RevId: 431197621
Change-Id: Ie9b2ae070abd4ffd11808a61d87169da6eb880bd