Fix TransformedArrayNDIterable support for inner index array dimensions
This fixes https://github.com/google/tensorstore/issues/31.

Thanks to @perlman for an excellent test case that made it very easy
to identify the bug.

This bug was missed by the existing unit tests because normally the
block size is chosen to be fairly large, e.g. 2048, and consequently
none of the previous test cases that exercised that particular code
path of nditerable_transformed_array.cc had an inner dimension
exceeding the block size.

To address that limitation, this change also adds a debugging option
`nditerable_test_unit_block_size` which allows testing with a block
size of 1.

That option uncovered another bug in the downsample driver for the
case that the block size is smaller than the downsample factor for the
inner dimension.  This change also fixes that bug.

PiperOrigin-RevId: 431496400
Change-Id: Ic040d1bb4806c2a9b365113a1b4c560287706d02