Change boolean metafunctions to variable template form and upgrade zlib
This also upgrades the zlib dependency to fix the build.

The previous zlib version has unfortunately been removed from the
upstream server.  It was removed because it contains a vulnerability,
but that vulnerability does not apply to TensorStore's use of the
library (nor to typical use of the library in general).

PiperOrigin-RevId: 437848697
Change-Id: I3a48dce0960f85a3c495b48ee0f143b53a80c84f