Beginning a bazel_to_cmake.py script to generate CMakeLists.txt
This script begins to generate CMakeLists.txt files for tensorstore.

The basic strategy is to encode cmake information into the BUILD and
workspace.bzl files, and then use this script to autogenerate CMakeLists.txt
files with the indicated  targets.

This is very much a work in progress, and not yet completed.

PiperOrigin-RevId: 438343528
Change-Id: I5804c837d51006ca13cc69b628913f0186cb088b