Add OpenMode::assume_metadata
This option allows redundant I/O to be avoided when creating/opening
TensorStores from multiple machines at the same time, at the cost of
reduced validation/safety.

This also fixes a bug in kvs_backed_chunk_driver.cc: previously, when
performing a transactional create operation, the create could
erroneously fail due to stale cached data, because no read request was
issued as part of the create operation (prior to commit).  With this
change, a read request is issued as part of the transactional create
operation, which avoids this problem.

PiperOrigin-RevId: 457526173
Change-Id: Ic640ee215537301ffe8b37e1aa3d3c75ba2c8c9b