Properly hide symbols in the Python extension _tensorstore.so
Due to a typo in the starlark macro, the version script intended to
restrict symbol visibility was not actually included.  The version
script should reduce the risk of symbol collisions when using system
libraries.

As version scripts are not supported on macOS, the macOS-specific
`-exported_symbol` option is used instead.

Additionally, this change modifies `setup.py` to specify
`-fvisibility=hidden` when building the Python extension.

PiperOrigin-RevId: 464124003
Change-Id: I79b97c995035b294b659f52c4825335e2b8e25fa