Use anonymous GCS requests when on GCE with no default service account
Previously, when running on GCE with no default service account
available (such as on Colab), the GceAuthProvider was used but then
would fail to generate a token, preventing a request from being made.

This commit changes the Google auth provider logic to check that a
default service account is available before choosing to use
GceAuthProvider.

Consequently, when running on colab, anonymous GCS requests are used
by default.

NOKEYCHECK

PiperOrigin-RevId: 466405536
Change-Id: I969141ca99512426c44e448f5cdba1312809000c