Avoid stale data from GCS
Previously, the GCS kvstore used cache-control request headers to
attempt to ensure stale data was not received from GCS.  However, GCS
does not appear to respect cache-control request headers.  With this
change, a unique query parameter is included in every request, as that
is known to be effective for bypassing GCS caching.

PiperOrigin-RevId: 470280903
Change-Id: Ib77dc4fbde86f9205c474e9177281cffcdcf20bb