Fix serialization (pickling) of neuroglancer_precomputed driver
Previously, the OpenConstraints metadata field was serialized as
`::nlohmann::json::object_t` but deserialized as `::nlohmann::json::`,
leading to incorrect data being loaded, and then an exception being
thrown due to the (error-prone) implicit conversion from
`::nlohmann::json` to `::nlohmann::json::object_t` that is possible
because `JSON_USE_IMPLICIT_CONVERSIONS` is enabled by default.

This commit fixes the specific bug, but in the future we will ideally
be able to build with `JSON_USE_IMPLICIT_CONVERSIONS=0` to better
detect issues like this with the compiler.

PiperOrigin-RevId: 475728633
Change-Id: I1734963920efc7fc78d52c0b00c19e36bae493e4