Improve GCS Rate Limiting for better distributed use.
Implement a rate limiting algorithm similar to the leaky-bucket for GCS. The idea is that instead of the strictly parallelism-based rate limiting, we can use an actual rate-limiting algorithm to slow down individual clients.

So when spinning up a lot of clients the spec could include, for example:

"experimental_gcs_rate_limiter" : {
  "read_rate": 5000,
  "write_rate": 1000,
  "doubling_time": "20min",
}

"5000" and "1000", of course, should be scaled by the number of clients.

See https://cloud.google.com/storage/docs/request-rate#ramp-up

PiperOrigin-RevId: 484340747
Change-Id: Ifebb948fa1089883ffc0a4d2e601033ed8c8c5b1