Allow {Result,Future}::status() to be called in success state
This provides uniformity with `absl::StatusOr` and eliminates a common
source of bugs.

The existing `tensorstore::GetStatus` function is no longer needed in
most cases, but may still be used in generic code that needs to
support `absl::Status`, `Result`, and `Future` uniformly.

Fixes https://github.com/google/tensorstore/issues/56

This also simplifies the definition of `Result`, and optimizes the
representation of `Result<void>`.

PiperOrigin-RevId: 484651697
Change-Id: Ib746431e292511e36b0fb644c304e66860069b0d