Add cc_grpc_library bazel build rules.
This variant is similar to the one which ships in grpc, except that it omits the proto_library() build and adds a namespace option for the generated code. It uses
the grpc generate_cc internal call to generate the grpc proto.

Also adds some (still incomplete) support to bazel_to_cmake to generate grpc proto services.

In the process fixes a few errors and missing codepaths in bazel_to_cmake.

PiperOrigin-RevId: 489595894
Change-Id: I516edc45159383e94e9f09c0fb4d36cb0ec13682